Add files via upload
In this game we have two players player1 and player2 when the game starts both the players start moving opposite sides if they colloid with each other or the “Head” of player one touch the other player the game is over so in this way it goes with the player two as well. Finally outcome shows as if player1 wins it shows player1 won same with player two.